Moved NounsToken functions to their own file.
